requirements.txt
discord.py[voice]==2.3.2
yt-dlp==2023.11.16
python-dotenv==1.0.0
aiohttp==3.8.5
asyncio==3.4.3